refactor(ids,ai): clear implementation from remaining mod.rs files (#477)
Sweep the four mod.rs files that still held a shared trait + core types
beyond the issue's original offender list, so every mod.rs is now
declarations + re-exports (routes/mod.rs keeps only its 3 shared
status-code glue helpers, which #477 permits):

- aifw-ids protocol/ -> types.rs (AppProto, Probe/ParseResult,
  StickyBuffers, ProtocolParser) + registry.rs (ProtocolRegistry)
- aifw-ids capture/ -> types.rs (CaptureConfig/RawPacket/CaptureStats +
  CaptureBackend) + factory.rs (create_capture)
- aifw-ids output/ -> pipeline.rs (AlertOutput + AlertPipeline)
- aifw-ai detectors/ -> registry.rs (Detector + run/default helpers)

Pure relocation; glob re-exports keep flat paths (protocol::AppProto,
capture::create_capture, output::AlertPipeline, detectors::Detector, etc.)
resolving. Code verified line-for-line identical apart from cfg-gated
backend imports in capture/factory.rs. cargo check + clippy --tests clean,
639 tests green.

Bump 5.97.16 -> 5.97.17.
