+++ aifw-ui/package.json
@@ -1,6 +1,6 @@
 {
   "name": "aifw-ui",
-  "version": "5.97.16",
+  "version": "5.97.17",
   "private": true,
   "scripts": {
     "dev": "next dev",
